refactor: 빌드에러로 인한 상수 위치 변경 및 일부 상수 별도 파일로 분리
