+++ src/app/constants/auth.ts
@@ -1,0 +1,1 @@
+export const AUTH_ACCESS_TOKEN = 'access_token';
